split composable to `presentation`
